+++ AWS-IoT-Arduino-Yun-Library/aws_iot_version.h
@@ -16,9 +16,9 @@
 #ifndef version_h
 #define version_h
 
-#define VERSION_MAJOR 1
-#define VERSION_MINOR 1
-#define VERSION_PATCH 2
+#define VERSION_MAJOR 2
+#define VERSION_MINOR 0
+#define VERSION_PATCH 0
 #define VERSION_TAG ""
 
 #endif